refactor(ui): extract Every Code queue
